pixelpro/backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic-settings==2.5.2

# Auth / Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AWS S3
boto3==1.35.0

# Task queue
celery==5.4.0
redis==5.1.0
flower==2.0.1

# AI / Vision
torch==2.4.1
torchvision==0.19.1
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4

# Real-ESRGAN
basicsr==1.4.2
realesrgan==0.3.0

# GFPGAN
gfpgan==1.3.8

# Background removal
rembg==2.0.57
onnxruntime-gpu==1.19.2  # use onnxruntime for CPU-only

# Payments
stripe==10.12.0

# Observability
prometheus-fastapi-instrumentator==7.0.0
sentry-sdk[fastapi]==2.16.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
gradio_client>=0.7.0
